requirements.txt
analytics-python==1.4.0
backoff==1.10.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
ffmpy==0.3.0
Flask==2.0.2
Flask-CacheBuster==1.0.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
fonttools==4.28.5
gradio==2.7.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
markdown2==2.4.2
MarkupSafe==2.0.1
matplotlib==3.5.1
monotonic==1.6
numpy==1.22.0
packaging==21.3
pandas==1.3.5
paramiko==2.9.1
Pillow==9.0.0
pycparser==2.21
pycryptodome==3.12.0
pydub==0.25.1
PyNaCl==1.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.0
six==1.16.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
